requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
bcrypt
pyotp
python-multipart
cryptography
python-dotenv
jinja2
pydantic[email]
qrcode[pil]
